SPECIALK keyboard patch.  Improves keyboard handling on older Apple IIe and IIc machines.
On some machines, there is a delay between when $C000 and $C010 react to a key press.
$C010 can indicate that key is down before $C000 has been updated with the actual ASCII
code of the new key.

If the game reads $C000 and $C010 while they are in this inconsistent state it causes the
last key to be repeated, rather than the key that was actually pressed.

The bug affects at least some IIe and IIc machines.  IIgs is not affected.

To fix, we use an unused memory location (ineditor) to record whether we should be tracking
a keypress or not.  It contains the state of the keyboard during the last pass of the game loop.
If a key wasn't down in the last loop, and $C000 doesn't contain a new key, then ignore $C010.