TOPCTRL.S is too big for track 4.
When CopyProtect=0 and ThreeFive=0 then the TOPCTRL binary is too big for track 4, by a single byte.

The missing byte causes the game to freeze when drinking the reversal potions on level 9. Reversing the screen forces a full-screen redraw. The redraw hook "develpatch" is at the very end of the file.

This bug was caused by change 9ac0241, which added 5 extra bytes of code to pass copy protection on side B and allow the final princess scene to run. Fixed, by removing some unneeded copy protection code.